chris-django-requirements.txt
Django==2.2.24
django-filter==2.4.0
djangorestframework==3.12.4
django-cors-middleware==1.5.0
psycopg2==2.8.6
python-swiftclient==3.12.0
django-storage-swift==1.2.19
celery==5.1.2
django-celery-beat==2.2.0
mod-wsgi==4.8.0
environs==9.3.2
python-pipchrisstoreclient==1.0.0
python-pfconclient==3.0.0
django-auth-ldap==4.1.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
collection-json==0.1.1
pudb>=2019.1
coverage==5.5
pylint==2.9.1
flake8==3.9.2
isort==5.9.1
boto3==1.24.66
django-storages==1.13.1
awscli==1.25.67
https://github.com/msbrogli/rpudb/archive/master.zip

